# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Showcase Portfolio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Member {
  id: number;
  name: string;
  role: string;
  /** @nullable */
  avatarUrl: string | null;
  /** @nullable */
  bio: string | null;
  /** @nullable */
  githubUrl: string | null;
  /** @nullable */
  linkedinUrl: string | null;
  /** @nullable */
  websiteUrl: string | null;
  createdAt: string;
}

export interface CreateMemberBody {
  name: string;
  role: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  live: "live",
  in_progress: "in_progress",
  archived: "archived",
} as const;

export interface Project {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  longDescription: string | null;
  tags: string[];
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  externalUrl: string | null;
  /** @nullable */
  githubUrl: string | null;
  status: ProjectStatus;
  featured: boolean;
  sortOrder: number;
  memberIds: number[];
  members: Member[];
  createdAt: string;
}

export type CreateProjectBodyStatus =
  (typeof CreateProjectBodyStatus)[keyof typeof CreateProjectBodyStatus];

export const CreateProjectBodyStatus = {
  live: "live",
  in_progress: "in_progress",
  archived: "archived",
} as const;

export interface CreateProjectBody {
  title: string;
  description: string;
  /** @nullable */
  longDescription?: string | null;
  tags?: string[];
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  externalUrl?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  status: CreateProjectBodyStatus;
  featured: boolean;
  sortOrder: number;
  memberIds?: number[];
}

export type UpdateProjectBodyStatus =
  (typeof UpdateProjectBodyStatus)[keyof typeof UpdateProjectBodyStatus];

export const UpdateProjectBodyStatus = {
  live: "live",
  in_progress: "in_progress",
  archived: "archived",
} as const;

export interface UpdateProjectBody {
  title?: string;
  description?: string;
  /** @nullable */
  longDescription?: string | null;
  tags?: string[];
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  externalUrl?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  status?: UpdateProjectBodyStatus;
  featured?: boolean;
  sortOrder?: number;
  memberIds?: number[];
}

export interface Summary {
  totalProjects: number;
  totalMembers: number;
  liveProjects: number;
  inProgressProjects: number;
  featuredCount: number;
}