# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Showcase Portfolio API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns all showcase projects in order
 * @summary List all projects
 */
export const GetProjectsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  longDescription: zod.string().nullable(),
  tags: zod.array(zod.string()),
  imageUrl: zod.string().nullable(),
  externalUrl: zod.string().nullable(),
  githubUrl: zod.string().nullable(),
  status: zod.enum(["live", "in_progress", "archived"]),
  featured: zod.boolean(),
  sortOrder: zod.number(),
  memberIds: zod.array(zod.number()),
  members: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      role: zod.string(),
      avatarUrl: zod.string().nullable(),
      bio: zod.string().nullable(),
      githubUrl: zod.string().nullable(),
      linkedinUrl: zod.string().nullable(),
      websiteUrl: zod.string().nullable(),
      createdAt: zod.string(),
    }),
  ),
  createdAt: zod.string(),
});
export const GetProjectsResponse = zod.array(GetProjectsResponseItem);

/**
 * @summary Create a new project
 */
export const CreateProjectBody = zod.object({
  title: zod.string(),
  description: zod.string(),
  longDescription: zod.string().nullish(),
  tags: zod.array(zod.string()).optional(),
  imageUrl: zod.string().nullish(),
  externalUrl: zod.string().nullish(),
  githubUrl: zod.string().nullish(),
  status: zod.enum(["live", "in_progress", "archived"]),
  featured: zod.boolean(),
  sortOrder: zod.number(),
  memberIds: zod.array(zod.number()).optional(),
});

/**
 * @summary Get a single project
 */
export const GetProjectParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProjectResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  longDescription: zod.string().nullable(),
  tags: zod.array(zod.string()),
  imageUrl: zod.string().nullable(),
  externalUrl: zod.string().nullable(),
  githubUrl: zod.string().nullable(),
  status: zod.enum(["live", "in_progress", "archived"]),
  featured: zod.boolean(),
  sortOrder: zod.number(),
  memberIds: zod.array(zod.number()),
  members: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      role: zod.string(),
      avatarUrl: zod.string().nullable(),
      bio: zod.string().nullable(),
      githubUrl: zod.string().nullable(),
      linkedinUrl: zod.string().nullable(),
      websiteUrl: zod.string().nullable(),
      createdAt: zod.string(),
    }),
  ),
  createdAt: zod.string(),
});

/**
 * @summary Update a project
 */
export const UpdateProjectParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateProjectBody = zod.object({
  title: zod.string().optional(),
  description: zod.string().optional(),
  longDescription: zod.string().nullish(),
  tags: zod.array(zod.string()).optional(),
  imageUrl: zod.string().nullish(),
  externalUrl: zod.string().nullish(),
  githubUrl: zod.string().nullish(),
  status: zod.enum(["live", "in_progress", "archived"]).optional(),
  featured: zod.boolean().optional(),
  sortOrder: zod.number().optional(),
  memberIds: zod.array(zod.number()).optional(),
});

export const UpdateProjectResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  longDescription: zod.string().nullable(),
  tags: zod.array(zod.string()),
  imageUrl: zod.string().nullable(),
  externalUrl: zod.string().nullable(),
  githubUrl: zod.string().nullable(),
  status: zod.enum(["live", "in_progress", "archived"]),
  featured: zod.boolean(),
  sortOrder: zod.number(),
  memberIds: zod.array(zod.number()),
  members: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      role: zod.string(),
      avatarUrl: zod.string().nullable(),
      bio: zod.string().nullable(),
      githubUrl: zod.string().nullable(),
      linkedinUrl: zod.string().nullable(),
      websiteUrl: zod.string().nullable(),
      createdAt: zod.string(),
    }),
  ),
  createdAt: zod.string(),
});

/**
 * @summary Delete a project
 */
export const DeleteProjectParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all team members
 */
export const GetMembersResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  role: zod.string(),
  avatarUrl: zod.string().nullable(),
  bio: zod.string().nullable(),
  githubUrl: zod.string().nullable(),
  linkedinUrl: zod.string().nullable(),
  websiteUrl: zod.string().nullable(),
  createdAt: zod.string(),
});
export const GetMembersResponse = zod.array(GetMembersResponseItem);

/**
 * @summary Create a team member
 */
export const CreateMemberBody = zod.object({
  name: zod.string(),
  role: zod.string(),
  avatarUrl: zod.string().nullish(),
  bio: zod.string().nullish(),
  githubUrl: zod.string().nullish(),
  linkedinUrl: zod.string().nullish(),
  websiteUrl: zod.string().nullish(),
});

/**
 * @summary Get a single member
 */
export const GetMemberParams = zod.object({
  id: zod.coerce.number(),
});

export const GetMemberResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  role: zod.string(),
  avatarUrl: zod.string().nullable(),
  bio: zod.string().nullable(),
  githubUrl: zod.string().nullable(),
  linkedinUrl: zod.string().nullable(),
  websiteUrl: zod.string().nullable(),
  createdAt: zod.string(),
});

/**
 * Returns total counts and highlights for the portfolio
 * @summary Get portfolio summary stats
 */
export const GetSummaryResponse = zod.object({
  totalProjects: zod.number(),
  totalMembers: zod.number(),
  liveProjects: zod.number(),
  inProgressProjects: zod.number(),
  featuredCount: zod.number(),
});

/**
 * @summary Get featured/highlighted projects
 */
export const GetFeaturedProjectsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  longDescription: zod.string().nullable(),
  tags: zod.array(zod.string()),
  imageUrl: zod.string().nullable(),
  externalUrl: zod.string().nullable(),
  githubUrl: zod.string().nullable(),
  status: zod.enum(["live", "in_progress", "archived"]),
  featured: zod.boolean(),
  sortOrder: zod.number(),
  memberIds: zod.array(zod.number()),
  members: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      role: zod.string(),
      avatarUrl: zod.string().nullable(),
      bio: zod.string().nullable(),
      githubUrl: zod.string().nullable(),
      linkedinUrl: zod.string().nullable(),
      websiteUrl: zod.string().nullable(),
      createdAt: zod.string(),
    }),
  ),
  createdAt: zod.string(),
});
export const GetFeaturedProjectsResponse = zod.array(
  GetFeaturedProjectsResponseItem,
);