# lib/api-zod/src/generated/types/createProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Showcase Portfolio API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateProjectBodyStatus } from "./createProjectBodyStatus";

export interface CreateProjectBody {
  title: string;
  description: string;
  /** @nullable */
  longDescription?: string | null;
  tags?: string[];
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  externalUrl?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  status: CreateProjectBodyStatus;
  featured: boolean;
  sortOrder: number;
  memberIds?: number[];
}