# lib/api-zod/src/generated/types/member.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Showcase Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface Member {
  id: number;
  name: string;
  role: string;
  /** @nullable */
  avatarUrl: string | null;
  /** @nullable */
  bio: string | null;
  /** @nullable */
  githubUrl: string | null;
  /** @nullable */
  linkedinUrl: string | null;
  /** @nullable */
  websiteUrl: string | null;
  createdAt: string;
}