# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Showcase Portfolio API
 * OpenAPI spec version: 0.1.0
 */
import type { Member } from "./member";
import type { ProjectStatus } from "./projectStatus";

export interface Project {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  longDescription: string | null;
  tags: string[];
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  externalUrl: string | null;
  /** @nullable */
  githubUrl: string | null;
  status: ProjectStatus;
  featured: boolean;
  sortOrder: number;
  memberIds: number[];
  members: Member[];
  createdAt: string;
}