# lib/api-zod/src/generated/types/summary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Showcase Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface Summary {
  totalProjects: number;
  totalMembers: number;
  liveProjects: number;
  inProgressProjects: number;
  featuredCount: number;
}